feat: Add signal to save additional user data from Google login
This commit adds a signal in the `autenticacion` app to save additional user data when a user logs in with Google using Allauth. The signal is triggered when a user logs in and checks if the user has linked their social account. If a social account exists, the signal extracts data from the Google account and updates the user's profile with the name, profile picture, and birthdate if available.